refactor: 불필요한 핸들링 메서드 제거

+++ src/components/CartCount.tsx
@@ -16,7 +16,7 @@ export const CartCount = memo(
 
     const decreaseCount = () => {
       if (quantity === 1) {
-        handleDeleteCart();
+        handleDeleteCart && handleDeleteCart();
         return;
       }
